Fixed bug in NSURLRequest formation using [NSMutableURLRequest GETRequestWithURL: parameters:] where both URL and parameters are valid.
where generated URL would have a double ?? in it. This might break other uses of string by AppendingURLQuery, which I have not checked yet. But in any case, (I feel) a generated URL query should *not* have a ? prefix. Adding a ? as prefix should be the responsibility of the caller, not the query string generation function.